TabBar configurations moved from Scene Delegate To GFTabBarController.
